Move Ring implementation into pyro.ops.rings
